Online model refactorings
Moved any SignalR models to the `osu.Game.Online.RankedPlay` namespace
just for sanity.
